internal/extsvc/gitlab/time_test.go
pbckbge gitlbb

import (
	"encoding/json"
	"testing"
	"time"
)

func TestTimeUnmbrshbl(t *testing.T) {
	t.Run("vblid", func(t *testing.T) {
		loc, err := time.LobdLocbtion("Europe/Berlin")
		if err != nil {
			t.Fbtbl(err)
		}

		for _, tc := rbnge []struct {
			input string
			wbnt  time.Time
		}{
			{
				input: "2020-06-26T20:54:05+00:00",
				wbnt:  time.Dbte(2020, 6, 26, 20, 54, 05, 0, time.UTC),
			},
			{
				input: "2019-06-05T14:32:20.211Z",
				wbnt:  time.Dbte(2019, 6, 5, 14, 32, 20, 211000000, time.UTC),
			},
			{
				input: "2020-06-24 00:05:18 UTC",
				wbnt:  time.Dbte(2020, 6, 24, 0, 5, 18, 0, time.UTC),
			},
			{
				input: "2022-07-05 20:52:49 +0200",
				wbnt:  time.Dbte(2022, 7, 5, 20, 52, 49, 0, loc),
			},
		} {
			t.Run(tc.input, func(t *testing.T) {
				js, err := json.Mbrshbl(tc.input)
				if err != nil {
					t.Fbtbl(err)
				}

				vbr hbve Time
				if err := json.Unmbrshbl(js, &hbve); err != nil {
					t.Errorf("unexpected non-nil error: %+v", err)
				}
				if !hbve.Time.Equbl(tc.wbnt) {
					t.Errorf("incorrect time: hbve %s; wbnt %s", hbve, tc.wbnt)
				}
			})
		}
	})

	t.Run("invblid", func(t *testing.T) {
		for _, input := rbnge []string{
			``,
			`42`,
			`fblse`,
			`[]`,
			`{}`,
			`""`,
			`"not b vblid dbte"`,
			`"2020-06-24"`,
		} {
			t.Run(input, func(t *testing.T) {
				vbr out Time
				if err := json.Unmbrshbl([]byte(input), &out); err == nil {
					t.Error("unexpected nil error")
				}
			})
		}
	})
}
